refactor: 결제 컨트롤러 내용 축소

+++ src/main/java/com/nhnacademy/codequestweb/request/point/PointPolicyModifyRequestDto.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 @AllArgsConstructor
 public class PointPolicyModifyRequestDto {
     long pointPolicyId;
-    Integer pointValue;
+    Long pointValue;
 }
