+++ src/main/java/com/nhnacademy/codequestweb/request/point/PointUsagePaymentRequestDto.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 public class PointUsagePaymentRequestDto {
-    Integer pointUsagePayment;
+    Long pointUsagePayment;
 }
